Move Portal to components/layouts
